Conflicts:
	applications/plugins/Compliant/assembly/AssemblyVisitor.h
	applications/plugins/Registration/examples/ICPregistration_usingcontacts.scn
	applications/plugins/SofaSimpleGUI/SofaGL.cpp
	examples/Components/forcefield/DistanceGridForceField2.scn
	framework/sofa/core/objectmodel/Base.cpp
	modules/SofaBaseMechanics/MechanicalObject.inl
	modules/SofaBoundaryCondition/FixedConstraint.inl
	modules/SofaMiscForceField/PREquivalentStiffnessForceField.cpp
	modules/SofaMiscForceField/PREquivalentStiffnessForceField.h
	modules/SofaMiscForceField/PREquivalentStiffnessForceField.inl
	modules/SofaMiscMapping/DistanceMapping.inl
	modules/SofaObjectInteraction/PartialRigidificationForceField.cpp
	modules/SofaObjectInteraction/PartialRigidificationForceField.h
	modules/SofaObjectInteraction/PartialRigidificationForceField.inl

Former-commit-id: d40503407e87b71dcef8d1c23d08707da1f4f339

+++ applications/plugins/Compliant/Compliant_gui/CompliantAttachPerformer.inl
@@ -190,7 +190,7 @@ void CompliantAttachPerformer<DataTypes>::start()
 
     assert(pickedNode);
     pickedParticleIndex = picked.indexCollisionElement;
-    if (mstateCollision->getSize()<picked.indexCollisionElement+1)
+    if ((unsigned int)mstateCollision->getSize()<picked.indexCollisionElement+1)
         pickedParticleIndex = 0;
 
     //-------- Mouse manipulator
